# Use insertRule  in browser