--- a/.jest.web.json
+++ b/.jest.web.json
@@ -1,0 +1,6 @@
+{
+  "globals": {
+    "__DEV__": true
+  },
+  "testEnvironment": "jsdom"
+}