--- a/src/models/BrowserStyleSheet.js
+++ b/src/models/BrowserStyleSheet.js
@@ -143,25 +143,23 @@ class SpeedyBrowserTag implements Tag {
     // Retrieve the sheet for the new style tag
     const sheet = sheetForTag(newEl)
 
-    for (const componentId in this.components) {
-      if (this.components.hasOwnProperty(componentId)) {
-        const comp = this.components[componentId]
-        const { cssFromDOM } = comp
-        const rules = stringifyRules([cssFromDOM])
-        const rulesSize = rules.length
+    Object.keys(this.components).forEach(componentId => {
+      const comp = this.components[componentId]
+      const { cssFromDOM } = comp
+      const rules = stringifyRules([cssFromDOM])
+      const rulesSize = rules.length
 
-        let injectedRules = 0
-        for (let j = 0; j < rulesSize; j += 1) {
-          if (safeInsertRule(sheet, rules[j], sheet.cssRules.length)) {
-            injectedRules += 1
-          }
+      let injectedRules = 0
+      for (let j = 0; j < rulesSize; j += 1) {
+        if (safeInsertRule(sheet, rules[j], sheet.cssRules.length)) {
+          injectedRules += 1
         }
-
-        comp.componentIndex = this.componentSizes.length
-        comp.css = rules.join(' ')
-        this.componentSizes.push(injectedRules)
       }
-    }
+
+      comp.componentIndex = this.componentSizes.length
+      comp.css = rules.join(' ')
+      this.componentSizes.push(injectedRules)
+    })
   }
 
   getComponentIds() {