--- a/src/models/ComponentStyle.js
+++ b/src/models/ComponentStyle.js
@@ -59,7 +59,7 @@ export default (
       if (!StyleSheet.instance.hasInjectedComponent(this.componentId)) {
         const placeholder =
           process.env.NODE_ENV !== 'production' ? `.${componentId} {}` : ''
-        StyleSheet.instance.deferredInject(componentId, true, placeholder)
+        StyleSheet.instance.deferredInject(componentId, true, [placeholder])
       }
     }
 
@@ -93,7 +93,7 @@ export default (
         return name
       }
 
-      const css = `\n${stringifyRules(flatCSS, `.${name}`)}`
+      const css = stringifyRules(flatCSS, `.${name}`)
       // NOTE: this can only be set when we inject the class-name.
       // For some reason, presumably due to how css is stringifyRules behaves in
       // differently between client and server, styles break.