@probablyup 2017-10-04
I’m on vacation without my laptop until the 13th! If someone hasn’t by
then, will do.
On Wed, Oct 4, 2017 at 10:47 PM mxstbr-bot <notifications@github.com> wrote:

> Fails
> 🚫 Please include a CHANGELOG entry. You can find it at CHANGELOG.md
> <https://github.com/styled-components/styled-components/blob/master/CHANGELOG.md>
> Warnings
> ⚠️ There are library changes, but not tests. That's OK as long as you're
> refactoring existing code
>
> Generated by 🚫 dangerJS <https://github.com/danger/danger-js/>
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/styled-components/styled-components/pull/1208#issuecomment-334284588>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAiy1kIAPOZkkPpzU3SIdS3GbA4xVAw9ks5so-7-gaJpZM4PuNd4>
> .
>