@schwers 2017-10-04
@probablyup is your benchmark suite on github by anychance? Happy to run through it if so, or I can spin up some later this week