@mxstbr 2017-10-04
Yeah: https://github.com/styled-components/benchmark