@mxstbr 2017-10-05
Maybe this code from emotion by @tkh44 helps too, who also use stylis now and also use `insertRule`: https://github.com/emotion-js/emotion/blob/bc85789f9b21a9a7a51ec70bc4b1217af3f2d90b/packages/emotion/src/index.js#L42-L104