@tkh44 2017-10-05
@mitchellhamilton and I were discussing making the insertion plugin a stand-alone package. The goal would be for users to just be able to import the plugin, add it to stylis, and provide a `sheet.insert` callback. 

Would you be interested in this? If you are, I imagine we could have it published in the next few days.