@mxstbr 2017-10-05
> Would you be interested in this?

Hell yes that'd be great, would certainly save us a lot of time!