@schwers 2017-10-09
Thanks again @tkh44! I started using `stylis-plugin-emotion` in a fork of this PR.

https://github.com/styled-components/styled-components/compare/master...schwers:insert-rules-ast?expand=1

@mxstbr I'm stuck on some test errors that I can't quite figure out yet in ^^. I'll be offline the next couple of days, can probably do some fixups later this week.