@geelen 2017-10-09
This is great. I got stuck on edge cases with my investigation into stylis as well, I think there definitely could be a need to change something internally there to give us an extra plugin hook, but I'm not sure.

Good luck!