@thysultan 2017-10-09
I also made [rule-sheet](https://github.com/thysultan/stylis.js/tree/master/plugins/rule-sheet) to show how you could implement a plugin.

@geelen Where all the edge cases related to "at-rules"?