@schwers 2017-10-09
I think the errors I'm getting are more fundamental; I updated this branch to use `stylis-plugin-emotion` and checked in the failing output for anyone interested https://github.com/styled-components/styled-components/pull/1208/commits/2fc51c926edb2abba9e0432edab198165e1da2f2

Some errors seem like something fundamental I'm not understanding is broken. These I'm surprised about. At a high level my patch changes `stringifyRules` to always return an array of css rules instead of one string, and then calls `.join('')` on the server side, and when not using "speedy insert rule" mode.

```
    Expected value to equal:
      [{"paddingTop": 10}, {"opacity": 0.9}]
    Received:
      [64, {"opacity": 0.9}]
    
    Difference:
    
    - Expected
    + Received
    
    Array [
    +  64,
       Object {
    -    "paddingTop": 10,
    -  },
    -  Object {
         "opacity": 0.9,
       },
     ]
      
      at Object.<anonymous> (src/primitives/test/primitives.test.js:31:32)
      at Promise.resolve.then.el (node_modules/p-map/index.js:46:16)
      at process._tickCallback (internal/process/next_tick.js:103:7)

  ● primitives › extending › should combine styles of extending components

    expect(received).toEqual(expected)
    
    Expected value to equal:
      [{"opacity": 0.9}, undefined]
    Received:
      [65, undefined]
    
    Difference:
    
    - Expected
    + Received
    
    Array [
    -  Object {
    -    "opacity": 0.9,
    -  },
    +  65,
       undefined,
     ]
      
      at Object.<anonymous> (src/primitives/test/primitives.test.js:42:49)
```

There's some category of whitespace related errors such as. These I'm not too susprised about, getting the same output seems doable, but otherwise I think I'll need to update tests to ignore whitespacing
```
    - Snapshot
    + Received
    
    "<style type="text/css" data-styled-components="" data-styled-components-is-local="false">/* sc-component-id: sc-global-2303210225 */
     body{background: papayawhip;}
     </style><style type="text/css" data-styled-components="b" data-styled-components-is-local="true">/* sc-component-id: sc-a */
     .sc-a {}
    -
     .b{color: red;}
     </style>"

------

    Expected value to equal:
      "<style data-styled-components=\"a b c\" data-styled-components-is-local=\"true\">
              /* sc-component-id: ONE */
              .ONE {}
              .a { color: blue; }
              /* sc-component-id: TWO */
              .TWO {}
              .b { color: red; }
            
    /* sc-component-id: THREE */
    .THREE {}
    .c{color: green;}</style>"
    Received:
      "<style data-styled-components=\"a b c\" data-styled-components-is-local=\"true\">
              /* sc-component-id: ONE */
              .ONE {}
              .a { color: blue; }
              /* sc-component-id: TWO */
              .TWO {}
              .b { color: red; }
            
    /* sc-component-id: THREE */
    .THREE {}.c{color: green;}</style>"
    ```