@philpl 2017-10-09
@schwers Am I understanding correctly that the `stylis-plugin-emotion` plugin parses everything down to objects? That'd be at least what I expect from their architecture.

Won't it be easier to use @thysultan's example?

> I also made rule-sheet to show how you could implement a plugin.

At least, as far as I'm aware, stylis middleware already has some signals to signify rules and at-rules.