@schwers 2017-10-09
@philpl its parsing out to strings as far as I can tell. Taking a peak at @thysultan's example the implementations seem pretty similar