@philpl 2017-10-09
@schwers oh ok, sorry 😆  I misread the RN tests as web tests and was a little confused.