@tkh44 2017-10-09
@schwers if I can be of any assistance feel free to ping me. 

@mxstbr @thysultan There are bugs in the rule-sheet plugin with nested selectors insertion order.  I'm sure you could work them out, but they are there as of right now.

@philpl everything is done with strings. Objects are just turned into strings. It is much faster to use emotion with strings so its kind of unfortunate about those RN tests.