@thysultan 2017-10-09
@tkh44 Thanks... patched. Turns out the bug was with stylis not passing a depth arg to plugins 😄