@mitchellhamilton 2017-10-10
@thysultan I'm pretty sure `stylis-rule-sheet` still doesn't insert nested rules in the correct order. The `depth` argument isn't really that useful here, what really matters is the relationship between the parent selector and the child selector. Since the parent rule is processed after all of the child rules you need to maintain a list of the parent selectors that correspond to the children and once you find the parent rule you insert it before all the rules that have that parent and remove that selector from the parent selector queue.

[Here's an example with `stylis-rule-sheet` and `stylis-plugin-emotion` that shows where `stylis-rule-sheet` inserts rules in the incorrect order.](https://codesandbox.io/s/3r7033y7w5)