@thysultan 2017-10-10
@mitchellhamilton I see. It didn't take flat css into account, pushed a patch for that.