@mxstbr 2017-12-21
I published this as `styled-components@dev` and ran the `react-native-web` benchmark suite against the old version and this version. Results:

```
BEFORE:

Deep tree 122.14ms ±47.48ms
Wide tree 229.65ms ±71.58ms

AFTER:

Deep tree 114.01ms ±53.78ms
Wide tree 268.23ms ±221.53ms
```

:confused: