@mxstbr 2018-01-23
@necolas hot damn, you tell me that now?! :wink: I ran A-gambits [CSS-in-JS benchmarks](https://github.com/A-gambit/CSS-IN-JS-Benchmarks) instead, these were the results:

```
—v2.2.1—

  styled-components-v2-plugin-decouple-cell:
    - Rerender Duration: 1679.2210000000002 ms
    - Mount Duration: 746.0050000000001 ms

  styled-components-v2-plugin-inline-styles:
    - Rerender Duration: 51.04682 ms
    - Mount Duration: 54.02579999999998 ms

  styled-components-v2-plugin:
    - Rerender Duration: 2375.741 ms
    - Mount Duration: 987.3149999999999 ms

—@dev—

  styled-components-v2-plugin-decouple-cell:
    - Rerender Duration: 94.95263999999997 ms
    - Mount Duration: 117.93180000000002 ms

  styled-components-v2-plugin-inline-styles:
    - Rerender Duration: 49.18634000000001 ms
    - Mount Duration: 49.85859999999998 ms

  styled-components-v2-plugin:
    - Rerender Duration: 88.24652000000002 ms
    - Mount Duration: 93.33000000000007 ms
```

Ignoring the inline-styles result since that's irrelevant for this PR, LOOK AT THOSE NUMBERS 😍😍😍😍😍

- 1679ms => 95ms
- 746ms => 117ms
- 2375ms => 88ms
- 987ms => 93ms

Let's clean up this PR and ship it?!